# docs/swagger.yaml
basePath: /
definitions:
  domain.ActorInput:
    properties:
      birth_place:
        type: string
      birth_year:
        type: integer
      language:
        type: string
      name:
        type: string
      rest_year:
        type: integer
      sex:
        type: string
      surname:
        type: string
    type: object
  domain.SignInInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  domain.SignUpInput:
    properties:
      email:
        type: string
      nickname:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - nickname
    - password
    type: object
  domain.UpdateActorInfo:
    properties:
      language:
        type: string
      name:
        type: string
      rest_year:
        type: integer
      sex:
        type: string
      surname:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for HollywoodStars Application.
  title: Swagger HollywoodStars App API
  version: "1.0"
paths:
  /actors:
    get:
      consumes:
      - application/json
      description: get all actors info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Get all actors
      tags:
      - actor
    post:
      consumes:
      - application/json
      description: add actor info
      parameters:
      - description: actor's info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.ActorInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Add actor
      tags:
      - actor
  /actors/id:
    delete:
      consumes:
      - application/json
      description: Delete actor info by id
      parameters:
      - description: int valid
        in: query
        minimum: 1
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Delete actor by id
      tags:
      - actor
    get:
      consumes:
      - application/json
      description: Get actor info by id
      parameters:
      - description: int valid
        in: query
        minimum: 1
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Get actor by id
      tags:
      - actor
    put:
      consumes:
      - application/json
      description: Update actor info by id
      parameters:
      - description: int valid
        in: query
        minimum: 1
        name: id
        type: integer
      - description: new actor's info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateActorInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Update actor by id
      tags:
      - actor
  /auth/refresh:
    get:
      consumes:
      - application/json
      description: Refresh token
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Refresh
      tags:
      - user
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Login in system
      parameters:
      - description: user's info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
        default:
          description: ""
          schema: {}
      summary: SignIn
      tags:
      - user
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Registration in system
      parameters:
      - description: user's info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      summary: SignUp
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"